--- blocks/before-after/block.json
@@ -1,50 +1,0 @@
-{
-	"$schema": "https://schemas.wp.org/trunk/block.json",
-	"apiVersion": 3,
-	"name": "frontblocks/before-after",
-	"version": "1.0.0",
-	"title": "Before After",
-	"category": "media",
-	"icon": "image-flip-horizontal",
-	"description": "Compare two images with a draggable before/after slider.",
-	"keywords": [ "before", "after", "comparison", "slider", "image" ],
-	"textdomain": "frontblocks",
-	"attributes": {
-		"beforeImageId": {
-			"type": "integer"
-		},
-		"beforeImageUrl": {
-			"type": "string",
-			"default": ""
-		},
-		"afterImageId": {
-			"type": "integer"
-		},
-		"afterImageUrl": {
-			"type": "string",
-			"default": ""
-		},
-		"beforeLabel": {
-			"type": "string",
-			"default": "Before"
-		},
-		"afterLabel": {
-			"type": "string",
-			"default": "After"
-		},
-		"initialPosition": {
-			"type": "number",
-			"default": 50
-		},
-		"align": {
-			"type": "string"
-		}
-	},
-	"supports": {
-		"html": false,
-		"align": [ "wide", "full" ]
-	},
-	"editorScript": "frontblocks-before-after-editor",
-	"viewScript": "frontblocks-before-after-view",
-	"style": "frontblocks-before-after-style"
-}
